code refactoring try numba 2
