+++ SimpleWeather/Extensions/Date + Extension.swift
@@ -1,0 +1,9 @@
+//
+//  Date + Extension.swift
+//  SimpleWeather
+//
+//  Created by Alexander on 24.06.2020.
+//  Copyright © 2020 Alexander Team. All rights reserved.
+//
+
+import Foundation
